Remove params.json
The file says it's used for regeneration, but I don't see how or why - all the
content's in the html file, and I'm not going to be regenerating that from a
json file.  I'm actually going to be modifying it a bunch to make it into a
template that supports posts, so that wouldn't work, anyways.